Revert some C++ code to simplify
